# backend/API/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          The HTTP status code of the response.
          example: 400
        type: integer
      error:
        description: |-
          The standard HTTP status text for the code.
          example: Bad Request
        type: string
      message:
        description: |-
          A user-friendly message describing the specific issue.
          example: Invalid request body or validation failed.
        type: string
    type: object
...
  models.FollowRelationship:
    properties:
      created_at:
        description: When the relationship was created.
        type: string
      followee_id:
        description: The ID of the user being followed.
        type: string
      follower_id:
        description: The ID of the user who follows.
        type: string
      status:
        description: 'Relationship status: pending, accepted, blocked.'
        type: string
      updated_at:
        description: When the relationship was last updated.
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the social-network application.
  title: Social-network Application API
  version: "1.0"
paths:
...
  /api/v1/follow:
    post:
      consumes:
      - application/json
      description: Creates an accepted follow relationship when both the follower
        and followee are public profiles.
      parameters:
      - description: Follow target {\
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FollowRelationship'
        "400":
          description: Invalid request or self-follow
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Only public profiles are supported
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Already following
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Follow a public user
      tags:
      - Follow
  /api/v1/follow/accept/{id}:
    put:
      description: Accepts a pending follow request for the authenticated user.
      parameters:
      - description: Follower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: accepted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing follower ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Follow request not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Accept a follow request
      tags:
      - Follow
  /api/v1/follow/delete/{id}:
    delete:
      description: Removes an accepted or pending follow relationship for the authenticated
        user.
      parameters:
      - description: Followee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: unfollowed'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing followee ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Follow relationship not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Unfollow a user
      tags:
      - Follow
  /api/v1/follow/requests:
    get:
      description: Returns pending and accepted follow requests for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.FollowRelationship'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get follow requests
      tags:
      - Follow
  /api/v1/follow/requests/pending:
    get:
      description: Returns pending follow requests for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.FollowRelationship'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get pending follow requests
      tags:
      - Follow
  /api/v1/followers:
    get:
      description: Returns accepted followers for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.FollowRelationship'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get followers
      tags:
      - Follow
  /api/v1/followers/delete/{id}:
    delete:
      description: Removes an accepted or pending follower relationship for the authenticated
        user.
      parameters:
      - description: Follower ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: removed'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing follower ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Follow relationship not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Remove a follower
      tags:
      - Follow
...
swagger: "2.0"